꼬인 거 해결
